+++ Cargo.toml
@@ -60,3 +60,4 @@ clippy = { version = "0.*", optional = true }
 
 [dev-dependencies]
 rand = "*"
+quickcheck = "0.3.1"
